# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Channel.ID": {
				"type": "string",
				"pattern": "^[a-f0-9]+$"
			},
			"Timestamp": {
				"type": "integer",
				"format": "int64"
			},
			"User.ID": {
				"type": "string",
				"example": "wangcai",
				"pattern": "^[a-z][a-z0-9]*$"
			},
			"Channel.Desc": {
				"properties": {
					"cid": {
						"$ref": "#/components/schemas/Channel.ID"
					},
					"createdAt": {
						"$ref": "#/components/schemas/Timestamp"
					},
					"creator": {
						"$ref": "#/components/schemas/User.ID"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"cid",
					"createdAt",
					"creator",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Channel.Create": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 30
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User.ChannelListItem": {
				"properties": {
					"cid": {
						"$ref": "#/components/schemas/Channel.ID"
					}
				},
				"required": [
					"cid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Channel.UpdateDesc": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 30
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Channel.MemberListItem": {
				"properties": {
					"uid": {
						"$ref": "#/components/schemas/User.ID"
					}
				},
				"required": [
					"uid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Int": {
				"type": "integer",
				"format": "int32"
			},
			"Message.PostResp": {
				"properties": {
					"i": {
						"$ref": "#/components/schemas/Int"
					}
				},
				"required": [
					"i"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message.To": {
				"type": "string",
				"pattern": "^(@[a-z][a-z0-9]*|[a-f0-9]+)$"
			},
			"Long": {
				"type": "integer",
				"format": "int64"
			},
			"Message.Post": {
				"properties": {
					"to": {
						"$ref": "#/components/schemas/Message.To"
					},
					"content": {
						"type": "string",
						"minLength": 1,
						"maxLength": 2048
					},
					"nonce": {
						"$ref": "#/components/schemas/Long"
					}
				},
				"required": [
					"to",
					"content",
					"nonce"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message.Common": {
				"properties": {
					"i": {
						"$ref": "#/components/schemas/Int"
					},
					"from": {
						"$ref": "#/components/schemas/User.ID"
					},
					"to": {
						"$ref": "#/components/schemas/Message.To"
					},
					"content": {
						"type": "string"
					},
					"ts": {
						"$ref": "#/components/schemas/Timestamp"
					}
				},
				"required": [
					"i",
					"from",
					"to",
					"content",
					"ts"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message.Event": {
				"properties": {
					"event": {
						"type": "string",
						"enum": [
							"join",
							"leave",
							"channel"
						]
					},
					"from": {
						"$ref": "#/components/schemas/User.ID"
					},
					"to": {
						"$ref": "#/components/schemas/Message.To"
					},
					"ts": {
						"$ref": "#/components/schemas/Timestamp"
					}
				},
				"required": [
					"event",
					"from",
					"to",
					"ts"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Message.Common.Exclude_keyofMessage.Common.to__": {
				"properties": {
					"i": {
						"type": "number",
						"format": "double"
					},
					"from": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"ts": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"i",
					"from",
					"content",
					"ts"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Message.Common.to_": {
				"$ref": "#/components/schemas/Pick_Message.Common.Exclude_keyofMessage.Common.to__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"User.Stats": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/Int"
					},
					"channel": {
						"$ref": "#/components/schemas/Int"
					},
					"online": {
						"$ref": "#/components/schemas/Int"
					}
				},
				"required": [
					"user",
					"channel",
					"online"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User.Desc": {
				"properties": {
					"uid": {
						"$ref": "#/components/schemas/User.ID"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"$ref": "#/components/schemas/Timestamp"
					}
				},
				"required": [
					"uid",
					"name",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User.CreateResp": {
				"properties": {
					"desc": {
						"$ref": "#/components/schemas/User.Desc"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"desc",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User.UserInputID": {
				"type": "string",
				"example": "WangCai",
				"pattern": "^[a-zA-Z][a-zA-Z0-9]*$",
				"minLength": 3,
				"maxLength": 16
			},
			"User.Password": {
				"type": "string",
				"example": "aStrongPassword",
				"minLength": 6,
				"maxLength": 20
			},
			"User.Create": {
				"properties": {
					"uid": {
						"$ref": "#/components/schemas/User.UserInputID"
					},
					"name": {
						"type": "string",
						"example": "旺财",
						"minLength": 1,
						"maxLength": 20
					},
					"password": {
						"$ref": "#/components/schemas/User.Password"
					}
				},
				"required": [
					"uid",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User.CreateToken": {
				"properties": {
					"uid": {
						"$ref": "#/components/schemas/User.UserInputID"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"uid",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.CreateResp.Exclude_keyofUser.CreateResp.desc__": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.CreateResp.desc_": {
				"$ref": "#/components/schemas/Pick_User.CreateResp.Exclude_keyofUser.CreateResp.desc__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"User.UpdatePassword": {
				"properties": {
					"password": {
						"type": "string"
					},
					"newPassword": {
						"$ref": "#/components/schemas/User.Password"
					}
				},
				"required": [
					"password",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User.UpdateDesc": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 30
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User.PeerListItem": {
				"properties": {
					"uid": {
						"$ref": "#/components/schemas/User.ID"
					}
				},
				"required": [
					"uid"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "xToken",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "chat-srv",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/channels": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Channel.Desc"
								}
							}
						}
					}
				},
				"summary": "Create a new channel.",
				"tags": [
					"Channel"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Channel.Create"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User.ChannelListItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get user's channel list.",
				"tags": [
					"Channel"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/channels/{cid}/desc": {
			"get": {
				"operationId": "GetDesc",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Channel.Desc"
								}
							}
						}
					}
				},
				"summary": "Get the channel's description.",
				"tags": [
					"Channel"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cid",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Channel.ID"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateDesc",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Channel.Desc"
								}
							}
						}
					}
				},
				"summary": "Update the channel's description.",
				"tags": [
					"Channel"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cid",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Channel.ID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Channel.UpdateDesc"
							}
						}
					}
				}
			}
		},
		"/channels/{cid}/members": {
			"get": {
				"operationId": "GetMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Channel.MemberListItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get the channel's member list.",
				"tags": [
					"Channel"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cid",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Channel.ID"
						}
					}
				]
			}
		},
		"/channels/{cid}": {
			"post": {
				"operationId": "Join",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Join the channel.",
				"tags": [
					"Channel"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cid",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Channel.ID"
						}
					}
				]
			},
			"delete": {
				"operationId": "Leave",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Leave the channel.",
				"tags": [
					"Channel"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cid",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Channel.ID"
						}
					}
				]
			}
		},
		"/messages": {
			"post": {
				"operationId": "PostMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Message.PostResp"
								}
							}
						}
					}
				},
				"summary": "Post message to some user or channel.",
				"tags": [
					"Message"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Message.Post"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"anyOf": [
											{
												"$ref": "#/components/schemas/Message.Common"
											},
											{
												"$ref": "#/components/schemas/Message.Event"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Pull new messages for the current user.",
				"tags": [
					"Message"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"description": "the time after which",
						"in": "query",
						"name": "since",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Timestamp"
						}
					},
					{
						"description": "max count of records returned. Defaults to 10.",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Int"
						}
					}
				]
			}
		},
		"/messages/{with}": {
			"get": {
				"operationId": "GetMessagesWith",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Omit_Message.Common.to_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Query history messages with some user or channel.",
				"tags": [
					"Message"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"description": "channel id or \\@\\{user id\\}",
						"in": "path",
						"name": "with",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Message.To"
						}
					},
					{
						"description": "the index start from 0",
						"in": "query",
						"name": "start",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Int"
						}
					},
					{
						"description": "max count of records returned. Defaults to 10.",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Int"
						}
					}
				]
			}
		},
		"/misc/stats": {
			"get": {
				"operationId": "GetStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User.Stats"
								}
							}
						}
					}
				},
				"summary": "Get system statistics.",
				"tags": [
					"Misc"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User.CreateResp"
								}
							}
						}
					}
				},
				"summary": "Register a new user.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User.Create"
							}
						}
					}
				}
			}
		},
		"/tokens": {
			"post": {
				"operationId": "CreateToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User.CreateResp"
								}
							}
						}
					}
				},
				"summary": "Create an access token, or say login.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User.CreateToken"
							}
						}
					}
				}
			}
		},
		"/users/{uid}/desc": {
			"get": {
				"operationId": "GetDescByID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User.Desc"
								}
							}
						}
					}
				},
				"summary": "Get description of the user specified by id.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uid",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/User.ID"
						}
					}
				]
			}
		},
		"/password": {
			"put": {
				"operationId": "UpdatePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Omit_User.CreateResp.desc_"
								}
							}
						}
					}
				},
				"summary": "Update the password of the current user.",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User.UpdatePassword"
							}
						}
					}
				}
			}
		},
		"/desc": {
			"get": {
				"operationId": "GetDesc",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User.Desc"
								}
							}
						}
					}
				},
				"summary": "Get description of the current user.",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateDesc",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User.Desc"
								}
							}
						}
					}
				},
				"summary": "Update description of the current user.",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User.UpdateDesc"
							}
						}
					}
				}
			}
		},
		"/peers": {
			"get": {
				"operationId": "GetPeers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User.PeerListItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get a list of users who ever privately chatted with the current user.",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tags": [
		{
			"name": "User",
			"description": "Operations about users"
		},
		{
			"name": "Channel",
			"description": "Operations for channels"
		},
		{
			"name": "Message",
			"description": "Operations for messages"
		}
	]
}